extracted ProductImageUpload component
#20
